Added Package parameter to DeployEnvironmentCommand (similar to Lambda tools)
