+++ src/Amazon.ElasticBeanstalk.Tools/Program.cs
@@ -15,7 +15,6 @@ namespace Amazon.ElasticBeanstalk.Tools
                 new List<ICommandInfo>()
                 {
                     new GroupHeaderInfo("Commands to deploy to Amazon Elastic Beanstalk:"),
-                    new CommandInfo<DeployPackageCommand>(DeployPackageCommand.COMMAND_NAME, DeployPackageCommand.COMMAND_DESCRIPTION, DeployPackageCommand.CommandOptions),
                     new CommandInfo<CreatePackageCommand>(CreatePackageCommand.COMMAND_NAME, CreatePackageCommand.COMMAND_DESCRIPTION, CreatePackageCommand.CommandOptions),
                     new CommandInfo<DeployEnvironmentCommand>(DeployEnvironmentCommand.COMMAND_NAME, DeployEnvironmentCommand.COMMAND_DESCRIPTION, DeployEnvironmentCommand.CommandOptions),
                     new CommandInfo<ListEnvironmentsCommand>(ListEnvironmentsCommand.COMMAND_NAME, ListEnvironmentsCommand.COMMAND_DESCRIPTION, ListEnvironmentsCommand.CommandOptions),
